Fix key change bug and timing issue
The keys would change even after the game had ended due to
with 'keyChangeTimer' being out of scope. This has been fixed by
declaring it first and defining it as an interval later.

The timing was weird before because 'changeKeys' would have a
large gap at the beginning of the game and then run normally. This was
fixed by moving the invocation after the call for 'setControls' and
calling it as is before running the interval.